added sitemap,robots,and changes structure of pages in order to have SSR pages
