+++ app/components/ThreeScene.tsx
@@ -1,6 +1,7 @@
+'use client'
 import { useEffect, useRef } from 'react';
 import * as THREE from 'three';
-import { GLTFLoader } from 'three/examples/jsm/Addons.js';
+import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader.js';
 interface GLTFResult {
   scene: THREE.Group;
   scenes: THREE.Group[];
