+++ app/robots.ts
@@ -1,0 +1,12 @@
+import type { MetadataRoute } from 'next'
+ 
+export default function robots(): MetadataRoute.Robots {
+  return {
+    rules: {
+      userAgent: '*',
+      allow: '/',
+      disallow: '/private/',
+    },
+    sitemap: 'https://www.desirecutieqb.com/sitemap.xml',
+  }
+}
